Fix conversions warnings in windows
git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@3452 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ trunk/openvibe-plugins/samples/trunc/src/ovpCScenarioImporterXML.cpp
@@ -205,7 +205,7 @@ boolean CScenarioImporterXML::doImport(IScenarioImporterContext& rScenarioImport
 		size_t l_iBufferLen=0;
 		size_t l_iFileLen;
 		l_oFile.seekg(0, ios::end);
-		l_iFileLen=l_oFile.tellg();
+		l_iFileLen=(size_t)l_oFile.tellg();
 		l_oFile.seekg(0, ios::beg);
 		while(l_iFileLen && l_bStatusOk)
 		{
